requirements.txt
amqp==5.3.1
billiard==4.2.1
boto3==1.36.2
botocore==1.36.2
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
jmespath==1.0.1
kombu==5.4.2
prompt_toolkit==3.0.49
pycurl==7.45.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.11.1
six==1.17.0
tzdata==2024.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
